feat: options for bow to not require arrows in inventory (#107)
Thanks for the commit and sorry for the long wait. We were kinda busy with university and just now started to look at the code again. We always appreciate help :). I will merge this and change the default value of enderiteBowWithInfinityNeedsArrow to true so that it doesn't change the default behavior. We will add you to the contributor list in the readme and on the download sites. Thanks again for the help.